HSEARCH-3384 Move environmnent-related code (services, beans) to a dedicated package
