+++ engine/src/main/java/org/hibernate/search/engine/common/impl/SearchIntegrationImpl.java
@@ -13,7 +13,7 @@ import org.hibernate.search.engine.backend.index.IndexManager;
 import org.hibernate.search.engine.backend.index.spi.IndexManagerImplementor;
 import org.hibernate.search.engine.backend.spi.BackendImplementor;
 import org.hibernate.search.engine.common.spi.SearchIntegration;
-import org.hibernate.search.engine.common.spi.BeanResolver;
+import org.hibernate.search.engine.environment.bean.spi.BeanResolver;
 import org.hibernate.search.engine.mapper.mapping.spi.MappingImplementor;
 import org.hibernate.search.engine.mapper.mapping.spi.MappingKey;
 import org.hibernate.search.util.SearchException;
