+++ engine/src/main/java/org/hibernate/search/engine/common/spi/SearchIntegrationBuilder.java
@@ -10,6 +10,7 @@ import java.util.Properties;
 
 import org.hibernate.search.engine.mapper.mapping.building.spi.MappingInitiator;
 import org.hibernate.search.engine.mapper.mapping.spi.MappingKey;
+import org.hibernate.search.engine.environment.bean.spi.BeanResolver;
 
 /**
  * @author Yoann Rodiere
