+++ engine/src/main/java/org/hibernate/search/engine/environment/bean/BeanProvider.java
@@ -4,10 +4,10 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.engine.common;
+package org.hibernate.search.engine.environment.bean;
 
 
-import org.hibernate.search.engine.common.spi.BeanResolver;
+import org.hibernate.search.engine.environment.bean.spi.BeanResolver;
 import org.hibernate.search.util.SearchException;
 
 /**
