+++ engine/src/main/java/org/hibernate/search/engine/environment/bean/BeanReference.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.engine.common;
+package org.hibernate.search.engine.environment.bean;
 
 /**
  * @author Yoann Rodiere
