+++ engine/src/main/java/org/hibernate/search/engine/environment/bean/spi/ImmutableBeanReference.java
@@ -4,12 +4,12 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.engine.common.spi;
+package org.hibernate.search.engine.environment.bean.spi;
 
 
 import java.util.StringJoiner;
 
-import org.hibernate.search.engine.common.BeanReference;
+import org.hibernate.search.engine.environment.bean.BeanReference;
 
 /**
  * @author Yoann Rodiere
